Fixed implementation of CRUD (not full)

+++ src/main/java/mif/vu/ikeea/Entity/Repository/TeamRepository.java
@@ -3,5 +3,5 @@ package mif.vu.ikeea.Entity.Repository;
 import mif.vu.ikeea.Entity.Team;
 import org.springframework.data.repository.CrudRepository;
 
-public interface TeamRepository extends CrudRepository<Team, Integer> {
+public interface TeamRepository extends CrudRepository<Team, Long> {
 }
